requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==5.2.5
django-ckeditor==6.7.3
django-heroku==0.3.1
django-js-asset==3.1.2
django-resized==1.0.3
django-widget-tweaks==1.5.0
gunicorn==23.0.0
packaging==25.0
pillow==11.3.0
psycopg2==2.9.10
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.9.0
